Safari-609.1.11.1/LayoutTests/fast/css/style-sharing-type-and-readonly.html

<!DOCTYPE html>

<script src="../../resources/js-test.js"></script>

<style>
  div[readonly="red"] {
    color: red;
  }
  div[type="blue"] {
    color: blue;
  }
</style>

<div black>Black</div>
<div readonly="red">Red</div>
<div type="blue">Blue</div>

<script>
description("Make sure special case style sharing for readonly and type attributes works.");

shouldBe('getComputedStyle(document.querySelector("[black]")).color','"rgb(0, 0, 0)"');
shouldBe('getComputedStyle(document.querySelector("[readonly]")).color','"rgb(255, 0, 0)"');
shouldBe('getComputedStyle(document.querySelector("[type]")).color','"rgb(0, 0, 255)"');
</script>
