third_party/WebKit/LayoutTests/fast/css/style-sharing-type-and-readonly-expected.txt

Make sure special case style sharing for readonly and type attributes works

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(document.querySelector("[black]")).color is "rgb(0, 0, 0)"
PASS getComputedStyle(document.querySelector("[readonly]")).color is "rgb(255, 0, 0)"
PASS getComputedStyle(document.querySelector("[type]")).color is "rgb(0, 0, 255)"
PASS successfullyParsed is true

TEST COMPLETE

Black
Red
Blue
